feat: Add core data functionality to standings view - updating data still requires a fix

+++ Barracudas/TeamsStandings+CoreDataProperties.swift
@@ -16,15 +16,15 @@ extension TeamsStandings {
         return NSFetchRequest<TeamsStandings>(entityName: "TeamsStandings")
     }
 
-    @NSManaged public var gb: Int16
+    @NSManaged public var gb: Int32
     @NSManaged public var group: String?
     @NSManaged public var league: String?
-    @NSManaged public var losses: Int16
+    @NSManaged public var losses: Int32
     @NSManaged public var pct: Double
-    @NSManaged public var rank: Int16
+    @NSManaged public var rank: Int32
     @NSManaged public var round: String?
     @NSManaged public var streak: String?
     @NSManaged public var teamname: String?
-    @NSManaged public var wins: Int16
+    @NSManaged public var wins: Int32
 
 }
